Cancel plDroid Krakow 2024
from the webite:

We are sorry to announce that Oskar, the originator and organiser of the PlSwift, PlDroid and NYSwifty conferences, passed away on 17 January.

This year's conferences are cancelled for the moment. We will try to refund your tickets as soon as possible. We also ask that you do not buy any more tickets.

On behalf of Oskar, we would like to thank you for your interest and attendance at the editions so far.